- [x] Build fixed new configuration.json added - [x] Refactored range models (RangeModels)
