+++ app/shared/com/ortb/controllers/traits/properties/ServiceControllerCorePropertiesContracts.scala
@@ -3,7 +3,8 @@ package shared.com.ortb.controllers.traits.properties
 import shared.com.ortb.controllers.implementations.WebApiResultImplementation
 import play.api.Logger
 import shared.com.ortb.manager.AppManager
-import shared.com.ortb.model.config.{ ConfigModel, DemandSidePlatformConfigurationModel, ServiceBaseModel, ServiceModel, ServicesModel }
+import shared.com.ortb.model.config.core.ServiceBaseModel
+import shared.com.ortb.model.config.{ ConfigModel, DemandSidePlatformConfigurationModel, ServiceModel, ServicesModel }
 import shared.io.loggers.DatabaseLogTracer
 
 trait ServiceControllerCorePropertiesContracts {
