+++ app/shared/com/ortb/model/config/DemandSideServiceModel.scala
@@ -1,10 +1,13 @@
 package shared.com.ortb.model.config
 
+import shared.com.ortb.model.config.core.ServiceBaseModel
+import shared.com.ortb.model.ranges.RangeDoubleModel
+
 case class DemandSideServiceModel(
   title : String,
   description : String,
   domain : String,
   port : Int,
   routing : Array[String],
-  biddingRandomRange : RangeDoubleModel
+  ownBiddingRandomRange : RangeDoubleModel
 ) extends ServiceBaseModel
