--- app/shared/com/ortb/model/config/DomainPortBaseModel.scala
@@ -1,6 +1,0 @@
-package shared.com.ortb.model.config
-
-abstract class DomainPortBaseModel {
-  val domain : String
-  val port : Int
-}
