+++ app/shared/com/ortb/model/config/DomainPortModel.scala
@@ -1,5 +1,7 @@
 package shared.com.ortb.model.config
 
+import shared.com.ortb.model.config.core.DomainPortBaseModel
+
 case class DomainPortModel(
   domain : String,
   port : Int
