--- app/shared/com/ortb/model/config/RangeFloatModel.scala
@@ -1,15 +1,0 @@
-package shared.com.ortb.model.config
-
-import scala.util.Random
-
-case class RangeFloatModel(
-  start : Float,
-  end : Float
-) extends RangeModel[Float] {
-  /**
-   * Get a number inclusive of start and exclusive of end
-   */
-  lazy override val randomInBetweenRange : Float = Random.between(
-    start,
-    end)
-}
