--- app/shared/com/ortb/model/config/RangeLongModel.scala
@@ -1,15 +1,0 @@
-package shared.com.ortb.model.config
-
-import scala.util.Random
-
-case class RangeLongModel(
-  start : Long,
-  end : Long
-) extends RangeModel[Long] {
-  /**
-   * Get a number inclusive of start and exclusive of end
-   */
-  lazy override val randomInBetweenRange : Long = Random.between(
-    start,
-    end)
-}
