--- app/shared/com/ortb/model/config/ServiceBaseModel.scala
@@ -1,7 +1,0 @@
-package shared.com.ortb.model.config
-
-abstract class ServiceBaseModel extends DomainPortBaseModel {
-  val title : String
-  val description : String
-  val routing : Array[String]
-}
