+++ app/shared/com/ortb/model/config/ServiceModel.scala
@@ -1,5 +1,7 @@
 package shared.com.ortb.model.config
 
+import shared.com.ortb.model.config.core.ServiceBaseModel
+
 case class ServiceModel(
   title : String,
   description : String,
