+++ app/shared/com/ortb/model/ranges/RangeDoubleModel.scala
@@ -5,11 +5,11 @@ import scala.util.Random
 case class RangeDoubleModel(
   start : Double,
   end : Double
-) {
+) extends RangeModel[Double] {
   /**
    * Get a number inclusive of start and exclusive of end
    */
-  lazy val randomInBetweenRange : Double = Random.between(
+  lazy override val randomInBetweenRange : Double = Random.between(
     start,
     end)
 }
