+++ app/shared/com/ortb/model/ranges/RangeFloatModel.scala
@@ -5,11 +5,11 @@ import scala.util.Random
 case class RangeFloatModel(
   start : Float,
   end : Float
-) {
+) extends RangeModel[Float] {
   /**
    * Get a number inclusive of start and exclusive of end
    */
-  lazy val randomInBetweenRange : Float = Random.between(
+  lazy override val randomInBetweenRange : Float = Random.between(
     start,
     end)
 }
