+++ app/shared/com/ortb/model/ranges/RangeLongModel.scala
@@ -5,11 +5,11 @@ import scala.util.Random
 case class RangeLongModel(
   start : Long,
   end : Long
-) {
+) extends RangeModel[Long] {
   /**
    * Get a number inclusive of start and exclusive of end
    */
-  lazy val randomInBetweenRange : Long = Random.between(
+  lazy override val randomInBetweenRange : Long = Random.between(
     start,
     end)
 }
